requirements.txt
admin-totals==1.0.1
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.2
django-chartjs==2.2.1
django-import-export==2.5.0
djangotoolbox==1.8.0
djongo==1.3.3
et-xmlfile==1.0.1
idna==2.10
iniconfig==1.1.1
MarkupPy==1.14
mysqlclient==2.0.2
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycoingecko==2.2.0
pymodm==0.4.3
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.2
pytz==2020.4
PyYAML==5.4.1
requests==2.25.0
soupsieve==2.0.1
sqlparse==0.2.4
tablib==3.0.0
toml==0.10.2
urllib3==1.26.2
xlrd==2.0.1
xlwt==1.3.0
